Assume LFS OID is SHA256 value if SHA256 field missing from LFS Info. This fix allows models that have been updated in their repo to be detected as hash mismatch and re-downloaded.
The Hugging face api does not have a field 'SHA' or 'SHA256' on LFS files.
The LFS oid is the SHA256 hash of the file so use that when the LFS SHA256 is missing, which it always will be.
Api example: https://huggingface.co/api/models/Qwen/Qwen3-0.6B/tree/main